requirements.txt
alembic==1.12.1
attrs==23.1.0
autopep8==2.0.4
blinker==1.6.3
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
